+++ Data/Map/Ordered/Strict.hs
@@ -32,13 +32,15 @@ module Data.Map.Ordered.Strict
 	, Index, findIndex, elemAt
 	-- * List conversions
 	, fromList, assocs, toAscList
+	-- * 'M.Map' conversion
+	, toMap
 	) where
 
 import Data.Foldable (foldl')
 import qualified Data.Map.Strict as M
 import Data.Map.Ordered.Internal
 	( OMap(..), empty, (<>|), (|<>), delete, filter, (\\), (|/\), (/\|), null, size
-	, member, notMember, lookup, findIndex, elemAt, assocs, toAscList, fromTV )
+	, member, notMember, lookup, findIndex, elemAt, assocs, toAscList, fromTV, toMap )
 import Data.Map.Util
 import Prelude hiding (filter, lookup, null)